# API Changes
- Portals are now loaded into instance form and now load directly from their associated file every 10 seconds.
- Some API related methods were tampered with to reflect the above change.

# Plugin Changes

- The liquid movement listening event is now utilized only through instance form portals and the initial block is the only point checked before the event is cancelled.
- The selection mode message was changed to better describe how to create the selection.
- The portal list command has a new message with a new key and sends detailed averaged region coordinate information alongside whether the portal is enabled or disabled (RED or GREEN).
- Fixed an issue with a cooldown placeholder.

+++ src/xzot1k/plugins/sp/api/objects/Portal.java
@@ -62,7 +62,7 @@ public class Portal {
      */
     public boolean delete() {
         File file = new File(getPluginInstance().getDataFolder(), "/portals/" + getPortalId() + ".yml");
-        if (file.exists()) {
+        if (file != null && file.exists()) {
             file.delete();
             return true;
         }
